pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "anarchobot"
version = "0.1.0"
description = "Small language model stack for Apple Silicon (MPS) using the Muon optimizer."
authors = [{ name = "AnarchoBot Maintainers" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.3.0",
    "datasets>=2.16.0",
    "sentencepiece>=0.1.99",
    "tqdm>=4.66.1",
    "numpy>=1.24.0",
    "pyyaml>=6.0.1",
    "accelerate>=0.24.1",
    "einops>=0.7.0",
]

[tool.setuptools.packages.find]
where = ["src"]
